requirements.txt
python-gitlab>=1.10.0
